124. Insertar usando TypeORM

+++ teslo-shop/src/products/entities/product.entity.ts
@@ -24,7 +24,7 @@ export class Product {
   @Column({ type: 'text' })
   public title: string;
 
-  @Column({ type: 'numeric', default: 0 })
+  @Column({ type: 'float', default: 0 })
   public price: number;
 
   @Column({ type: 'text', nullable: true })
